Add @Shared, split to encryptService, move responses
